Restructure how the defines are loaded
The code to allow custom defines never actually attempted to load custom defines.php files, modified this to actually do so.  In accomplishing this, the version define needed to be moved so it would always be defined by us.  The updated defines codes enables the code to be symlinked to different locations, but there's some cleanup necessary to accomplish this.